More bdaddr fixes

+++ include/bluegrass/bluetooth.hpp
@@ -63,9 +63,12 @@ namespace bluegrass {
 	std::ostream& operator<<(std::ostream& out, const bdaddr_t& ba)  
 	{
 		std::ios_base::fmtflags ff(out.flags());
-		out << std::hex << std::uppercase << ba.b[5] << ':' << ba.b[4] << ':' 
-			<< ba.b[3] << ':' << ba.b[2] << ':' << ba.b[1] << ':' << ba.b[0];
+		out << std::hex << std::uppercase 
+		<< (uint32_t) ba.b[5] << ':' << (uint32_t) ba.b[4] << ':' << (uint32_t) ba.b[3] 
+		<< ':' 
+		<< (uint32_t) ba.b[2] << ':' << (uint32_t) ba.b[1] << ':' << (uint32_t) ba.b[0];
 		out.flags(ff);
+		
 		return out;
 	}
